refactor(ai-sdk): split the OpenAI provider into providers/openai/
Move the grab-bag providers/openai.ts into a directory split by concern
(chat, prompt-cache, embeddings, images, tts, stt, files, vector-store,
config), with providers/openai.ts kept as a re-export barrel so every
import path is unchanged, including the OpenAI-compatible derivatives
that pull OpenAIAdapter from it. The duplicated sleep() helper now comes
from util/sleep.ts.

Pure code movement, no behavior change. The per-adapter getClient()
copies are moved as-is; deduping them is #972.

Refs #973
