+++ packages/ai-sdk/src/providers/openai/index.ts
@@ -1,0 +1,10 @@
+export type { OpenAIConfig } from './config.js'
+export { OpenAIProvider } from './provider.js'
+export {
+  OpenAIAdapter,
+  normalizeToolTranscript,
+  toOpenAIMessages,
+  toOpenAITools,
+  mapOpenAIFinishReason,
+} from './chat.js'
+export { buildPromptCacheKey } from './prompt-cache.js'
